requirements.txt
fastapi
uvicorn
SQLAlchemy
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
pydantic
pydantic[email]
email-validator
pytest
httpx
pytest-asyncio
httpx